+++ Backend/src/Services/School.AdminService/DataTransferObjects/AddressDto.cs
@@ -4,12 +4,12 @@ namespace School.AdminService.DataTransferObjects
 {
     public class AddressDto
     {
-        public int AddressId { get; set; }
+        public long AddressId { get; set; }
         public int DistrictsId { get; set; }
         public int ThanasId { get; set; }
         public string StreetDetails { get; set; }
         [Required]
         public string AddressType { get; set; }
-        public int PersonalInfoId { get; set; }
+        public long PersonalInfoId { get; set; }
     }
 }
